lots of refactoring
feat(repl): repl vsplit is 50% of window
feat(lua-utils): added lua-utils as a builtin library
feat(Bookmark): marks picker not opened by default

+++ lua/bootstrap.lua
@@ -15,9 +15,7 @@ local function get_lazy()
   vim.opt.rtp:prepend(lazypath)
 
   local ok = pcall(require, "lazy")
-  if not ok then
-    error "FATAL ERROR: Could not install lazy.nvim"
-  end
+  if not ok then error "FATAL ERROR: Could not install lazy.nvim" end
 end
 get_lazy()
 
@@ -25,11 +23,13 @@ local function add_paths()
   -- Enable support for nvim-local luarocks
   local data_dir = vim.fn.stdpath "data"
   local home_dir = os.getenv "HOME" .. "/.nvim"
-  package.cpath = package.path .. ";" .. data_dir .. "/luarocks/share/lua/5.1/?.so"
-  package.cpath = package.path .. ";" .. data_dir .. "/luarocks/lib/lua/5.1/?.so"
-  package.path = package.path .. ";" .. data_dir .. "/luarocks/share/lua/5.1/?.lua"
-  package.path = package.path .. ";" .. data_dir .. "/luarocks/share/lua/5.1/?/?.lua"
-  package.path = package.path .. ";" .. data_dir .. "/luarocks/share/lua/5.1/?/init.lua"
+  local dir = table.concat({ os.getenv "HOME", ".luarocks" }, "/")
+
+  package.cpath = package.cpath .. ";" .. dir .. "/share/lua/5.1/?.so"
+  package.cpath = package.cpath .. ";" .. dir .. "/lib/lua/5.1/?.so"
+  package.path = package.path .. ";" .. dir .. "/share/lua/5.1/?.lua"
+  package.path = package.path .. ";" .. dir .. "/share/lua/5.1/?/?.lua"
+  package.path = package.path .. ";" .. dir .. "/share/lua/5.1/?/init.lua"
 
   -- Enable support for user config in ~/.nvim
   package.path = package.path .. ";" .. home_dir .. "/lua/?.lua"
@@ -40,15 +40,16 @@ add_paths()
 
 local function install_rocks()
   -- Install missing luarocks
-  local dest = vim.fn.stdpath "data" .. "/luarocks"
   function install_luarock(rock, req)
     local _
     ok, _ = pcall(require, req)
-    local cmd = string.format("luarocks --lua-version 5.1 --tree %s install %s", dest, rock)
+    local cmd =
+      string.format("luarocks --local --lua-version 5.1 install %s", dest, rock)
     if not ok then
       print("Attempting to install luarock " .. rock)
 
-      vim.fn.system(cmd)
+      local out = vim.fn.system(cmd)
+      print(out)
 
       ok, _ = pcall(require, req)
       if not ok then
@@ -57,17 +58,17 @@ local function install_rocks()
     end
   end
 
-  install_luarock("lua-yaml", "yaml")
-  install_luarock("penlight", "pl.stringx")
   install_luarock("lualogging", "logging.file")
   install_luarock("lrexlib-pcre2", "rex_pcre2")
+
+  -- Yet to upload
+  -- install_luarock('lua-utils', 'utils')
 end
 install_rocks()
 
 local function create_globals(args)
   -- Make some global variables
   local log_path = vim.fn.stdpath "config" .. "/nvim.log"
-  yaml = require "yaml"
   path = require "pl.path"
   file = require "pl.file"
   dir = require "pl.dir"
@@ -75,13 +76,24 @@ local function create_globals(args)
   json = { encode = vim.fn.json_encode, decode = vim.fn.json_decode }
   logger = logging.file(log_path, "", "[%date] [%level]\n %message\n\n")
   user = {}
+  utils = require "lua-utils.utils"
+  array = require "lua-utils.array"
+  dict = require "lua-utils.dict"
+  class = require "lua-utils.class"
+  multimethod = require "lua-utils.multimethod"
+  exception = require "lua-utils.exception"
+  local validate = require "lua-utils.validate"
+  local fn = require "lua-utils.fn"
+  local types = require "lua-utils.types"
+  dict.each(validate, function(key, value) _G[key] = value end)
+  dict.each(fn, function(key, value) _G[key] = value end)
+  dict.each(types, function(key, value) _G[key] = value end)
+  str = require "lua-utils.str"
 
   -- Delete the old log
-  if path.exists(log_path) then
-    vim.fn.system("rm " .. log_path)
-  end
+  if path.exists(log_path) then vim.fn.system("rm " .. log_path) end
 end
 create_globals()
 
 -- My modifications to penlight class
-require "utils"
+-- require "core.utils"
